int to float

+++ app.py
@@ -1,8 +1,8 @@
 #calculadora
 
 # Receive two input from the user as numbers and one input as an operator
-num1 = int(input('type the first number: '))
-num2 = int(input('type the second number: '))
+num1 = float(input('type the first number: '))
+num2 = float(input('type the second number: '))
 operator = (input('type the operator: '))
 
 # Perform the operation based on the operator input
